requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
tinydb>=4.0.0
requests>=2.25.1
python-dotenv>=1.0.0
